Implemented API wbdb-name to get the db name from a datomic URI.
In order to keep the size of utils.clj down, I wanted to split
up the file but not burden consumers having to know which namespace
to get it form, so decided to split up the module.

A web search showed there are generally two approaches in the
community for doing this:
 1. Adding the potemkin library as dependency, and creating new
    namespaces.
 2. Use the same pattern as clojure.core, where sub-modules are
    loaded into the main namespace.

In this implementation, I chose approach 2.

Since utils_wbdb.clj isn't a namespace in it's own right, but simply
contributes towards pseudoace.utils, I've disabled eastwood linter
`:no-ns-form-found` in the lein :dev profile.